Conflicts:
	cfg/local_ontologies.txt
	cfg/ontologyTermsAndNamesCache.txt
	cfg/startSettings.txt
	src/semgen/stage/StageWorkbench.java
	src/semsim/annotation/Relation.java
	src/semsim/definitions/RDFNamespace.java
	src/semsim/definitions/SemSimRelations.java
	src/semsim/definitions/SemSimTypes.java

+++ src/semsim/model/collection/SemSimModel.java
@@ -42,6 +42,7 @@ import semsim.model.physical.object.PhysicalProperty;
 import semsim.model.physical.object.PhysicalPropertyinComposite;
 import semsim.model.physical.object.ReferencePhysicalEntity;
 import semsim.model.physical.object.ReferencePhysicalProcess;
+import semsim.reading.ModelAccessor;
 import semsim.utilities.SemSimCopy;
 import semsim.writing.SemSimOWLwriter;
 
@@ -85,7 +86,7 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 	
 	private String namespace;
 	private int sourceModelType;
-	private String sourcefilelocation;
+	private ModelAccessor sourcefilelocation;
 	private double semsimversion;
 	
 	// Model-specific data
@@ -123,7 +124,10 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 	private SemSimModel(SemSimModel ssmtocopy) {
 		super(ssmtocopy);
 		namespace = new String(ssmtocopy.namespace);
-		sourcefilelocation = new String(ssmtocopy.sourcefilelocation);
+		
+		if(ssmtocopy.sourcefilelocation != null)
+			sourcefilelocation = new ModelAccessor(ssmtocopy.sourcefilelocation);
+		
 		sourceModelType = ssmtocopy.sourceModelType;
 		semsimversion = ssmtocopy.semsimversion;
 		importCurationalMetadatafromModel(ssmtocopy, true);
@@ -388,8 +392,13 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 		
 	/**@return The parent FunctionalSubmodel for a MappableVariable.*/
 	public FunctionalSubmodel getParentFunctionalSubmodelForMappableVariable(MappableVariable var){
-		String compname = var.getName().substring(0, var.getName().lastIndexOf("."));
-		return (FunctionalSubmodel) getSubmodel(compname);
+		
+		if(var.getName().contains(".")){
+			String compname = var.getName().substring(0, var.getName().lastIndexOf("."));
+			
+			return (FunctionalSubmodel) getSubmodel(compname);
+		}
+		else return null;
 	}
 	
 	/**
@@ -744,11 +753,11 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 	 * @return The location of the raw computer source code associated with this model.
 	 */
 	
-	public String getLegacyCodeLocation() {
+	public ModelAccessor getLegacyCodeLocation() {
 		return sourcefilelocation;
 	}
 
-	public void setSourceFileLocation(String sourcefilelocation) {
+	public void setSourceFileLocation(ModelAccessor sourcefilelocation) {
 		this.sourcefilelocation = sourcefilelocation;
 	}
 	
@@ -809,7 +818,7 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 				}
 				
 				FunctionalSubmodel fs = getParentFunctionalSubmodelForMappableVariable(mapv);
-				fs.removeVariableEquationFromMathML(mapv);
+				if(fs != null) fs.removeVariableEquationFromMathML(mapv);
 			}
 			
 			for(Submodel sub : getSubmodels()){
@@ -857,9 +866,9 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 	}
 
 	/**
-	 * Specify which modeling language was used for the original version of the model.
+	 * Specify which format was used for the model's simulation source code.
 	 * See {@link ModelClassifier} constants. 
-	 * @param originalModelType An integer corresponding to the language of the original model code (see {@link ModelClassifier} ).
+	 * @param originalModelType An integer corresponding to the format of the original model code (see {@link ModelClassifier} ).
 	 */
 	public void setSourceModelType(int originalModelType) {
 		this.sourceModelType = originalModelType;
@@ -867,7 +876,7 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 
 	
 	/**
-	 * @return An integer representing the language of the original model code (see {@link ModelClassifier} ) 
+	 * @return An integer representing the format of the original model code (see {@link ModelClassifier} ) 
 	 * and associated constants.
 	 */
 	public int getSourceModelType() {
@@ -969,7 +978,7 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 		metadata.setAnnotationValue(Metadata.description, value);
 	}
 	
-	public double getSemsimversion() {
+	public double getSemSimVersion() {
 		return semsimversion;
 	}
 
@@ -977,7 +986,7 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 		this.semsimversion = semsimversion;
 	}
 	
-	public void setSemsimversion(String semsimversion) {
+	public void setSemSimVersion(String semsimversion) {
 		this.semsimversion = Double.valueOf(semsimversion);
 	}
